Implement GNURadio-compatible data types across all signal processing
GNURadio Signal Data Types 표준에 맞추어 모든 데이터 타입을 명시적으로 정의

주요 변경 사항:
1. config.py: BPSK constellation을 np.complex64로 변경
2. signal_generator.py:
   - 비트 데이터: uint8 타입 명시
   - 복소수 신호: complex64 타입 명시
   - 모든 BPSK 변조/복조에 타입 지정
3. adc.py:
   - Float 출력: float32 타입 (GNURadio Float)
   - Integer 출력: int16 타입 (GNURadio Short)
4. ber_calculator.py: STF 참조 비트를 uint8로 변경
5. simulation_utils.py: 복조 비트를 uint8로 변경

GNURadio 타입 매핑:
- Complex: np.complex64 (64비트 = float32 실수부 + float32 허수부)
- Float: np.float32 (32비트)
- Short: np.int16 (16비트)
- Byte: np.uint8 (8비트)

새 파일:
- GNURADIO_COMPATIBILITY.md: 상세한 호환성 문서
- gnuradio_screenshot.jpeg: GNURadio wiki 참조 스크린샷

이점:
- GNURadio와의 직접 연동 가능
- 메모리 효율성 향상 (float64→float32, int→uint8)
- 무선통신 시뮬레이션 표준 준수

+++ ber_calculator.py
@@ -24,7 +24,8 @@ class BERCalculator:
         self.signal_generator = SignalGenerator()
         
         # STF 참조 비트 시퀀스 (원본)
-        self.stf_reference_bits = np.array(STF_BITS)
+        # GNURadio 호환: uint8 타입
+        self.stf_reference_bits = np.array(STF_BITS, dtype=np.uint8)
         self.stf_length = len(self.stf_reference_bits)
         
         # BER 계산을 위한 슬라이딩 윈도우
